Make nexus-test-utils no longer depend on Nexus, for the most part.
Not actually. There is definitely still a dependency, but the "test
setup" part can be re-used with Nexus tests without causing any real
circular dependency issues.

See https://github.com/oxidecomputer/omicron/pull/1835#discussion_r1001936727
for an example of one such issue.

Frankly the rest of nexus-test-utils probably *should* continue to be
refactored such that there is no real dependency on Nexus; doing so
will reduce the risk of "mis-matched versions of Nexus" classes of
errors in the future.

This required the following changes:

- Move `nexus`'s config information into `common/src/nexus_config.rs`,
  so that it can be used by `nexus-test-utils`.
- Make `nexus-test-utils` operate on a generic version of `NexusServer`,
  exposed by a new crate named `nexus-test-interface`. This crate can
  contain any information used to operate Nexus during shared test
  setup, without actually requiring a full dependency on `nexus`.
- Note, adding this generic required making `ControlPlaneTestContext`
  generic, which ended up bubbling up to a lot of test setup functions,
  including the one used in the `#[nexus_test]` macro...
- ... So, make it possible to customize which version of `NexusServer`
  is used by `#[nexus_test]`. This means callers can supply their *own*
  version of `NexusServer`, either from their own crate (see: Nexus
  unit tests) or the version from the `omicron_nexus` crate (see: Nexus
  integration tests).
