+++ common/Cargo.toml
@@ -36,5 +36,6 @@ progenitor = { git = "https://github.com/oxidecomputer/progenitor" }
 
 [dev-dependencies]
 expectorate = "1.0.5"
+libc = "0.2.135"
 serde_urlencoded = "0.7.1"
 tokio = { version = "1.21", features = [ "test-util" ] }
