+++ nexus/db-model/src/ipv4net.rs
@@ -8,8 +8,8 @@ use diesel::pg::Pg;
 use diesel::serialize::{self, ToSql};
 use diesel::sql_types;
 use ipnetwork::IpNetwork;
-use nexus_defaults as defaults;
 use omicron_common::api::external;
+use omicron_common::nexus_config::NUM_INITIAL_RESERVED_IP_ADDRESSES;
 use std::net::Ipv4Addr;
 
 #[derive(Clone, Copy, Debug, PartialEq, AsExpression, FromSqlRow)]
@@ -26,7 +26,7 @@ impl Ipv4Net {
             && (
                 // First N addresses are reserved
                 self.iter()
-                    .take(defaults::NUM_INITIAL_RESERVED_IP_ADDRESSES)
+                    .take(NUM_INITIAL_RESERVED_IP_ADDRESSES)
                     .all(|this| this != addr)
             )
             && (
