+++ nexus/db-model/src/ipv6net.rs
@@ -8,8 +8,8 @@ use diesel::pg::Pg;
 use diesel::serialize::{self, ToSql};
 use diesel::sql_types;
 use ipnetwork::IpNetwork;
-use nexus_defaults as defaults;
 use omicron_common::api::external;
+use omicron_common::nexus_config::NUM_INITIAL_RESERVED_IP_ADDRESSES;
 use rand::{rngs::StdRng, SeedableRng};
 use std::net::Ipv6Addr;
 
@@ -77,7 +77,7 @@ impl Ipv6Net {
         self.contains(addr)
             && self
                 .iter()
-                .take(defaults::NUM_INITIAL_RESERVED_IP_ADDRESSES)
+                .take(NUM_INITIAL_RESERVED_IP_ADDRESSES)
                 .all(|this| this != addr)
     }
 }
