+++ nexus/src/app/vpc_subnet.rs
@@ -13,7 +13,6 @@ use crate::db::model::Name;
 use crate::db::model::VpcSubnet;
 use crate::db::queries::vpc_subnet::SubnetError;
 use crate::external_api::params;
-use nexus_defaults as defaults;
 use omicron_common::api::external;
 use omicron_common::api::external::CreateResult;
 use omicron_common::api::external::DataPageParams;
@@ -21,6 +20,7 @@ use omicron_common::api::external::DeleteResult;
 use omicron_common::api::external::ListResultVec;
 use omicron_common::api::external::LookupResult;
 use omicron_common::api::external::UpdateResult;
+use omicron_common::nexus_config::MIN_VPC_IPV4_SUBNET_PREFIX;
 use uuid::Uuid;
 
 impl super::Nexus {
@@ -48,7 +48,7 @@ impl super::Nexus {
                 "VPC Subnet IPv4 address ranges must be from a private range",
             ));
         }
-        if params.ipv4_block.prefix() < defaults::MIN_VPC_IPV4_SUBNET_PREFIX
+        if params.ipv4_block.prefix() < MIN_VPC_IPV4_SUBNET_PREFIX
             || params.ipv4_block.prefix()
                 > self.tunables.max_vpc_ipv4_subnet_prefix
         {
@@ -57,7 +57,7 @@ impl super::Nexus {
                     "VPC Subnet IPv4 address ranges must have prefix ",
                     "length between {} and {}, inclusive"
                 ),
-                defaults::MIN_VPC_IPV4_SUBNET_PREFIX,
+                MIN_VPC_IPV4_SUBNET_PREFIX,
                 self.tunables.max_vpc_ipv4_subnet_prefix,
             )));
         }
