+++ nexus/test-interface/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "nexus-test-interface"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[dependencies]
+async-trait = "0.1.56"
+dropshot = { git = "https://github.com/oxidecomputer/dropshot", branch = "main" , features = [ "usdt-probes" ] }
+omicron-common = { path = "../../common" }
+slog = { version = "2.7" }
