+++ nexus/test-utils-macros/Cargo.toml
@@ -7,5 +7,6 @@ edition = "2021"
 proc-macro = true
 
 [dependencies]
+proc-macro2 = { version = "1.0" }
 quote = "1.0"
 syn = { version="1.0", features=["full", "fold", "parsing"] }
