+++ nexus/tests/integration_tests/oximeter.rs
@@ -4,13 +4,15 @@
 
 //! Integration tests for oximeter collectors and producers.
 
-use nexus_test_utils::ControlPlaneTestContext;
 use nexus_test_utils_macros::nexus_test;
 use omicron_test_utils::dev::poll::{wait_for_condition, CondCheckError};
 use oximeter_db::DbWrite;
 use std::net;
 use std::time::Duration;
 use uuid::Uuid;
+
+type ControlPlaneTestContext =
+    nexus_test_utils::ControlPlaneTestContext<omicron_nexus::Server>;
 
 #[nexus_test]
 async fn test_oximeter_database_records(context: &ControlPlaneTestContext) {
@@ -62,8 +64,10 @@ async fn test_oximeter_database_records(context: &ControlPlaneTestContext) {
 
 #[tokio::test]
 async fn test_oximeter_reregistration() {
-    let mut context =
-        nexus_test_utils::test_setup("test_oximeter_reregistration").await;
+    let mut context = nexus_test_utils::test_setup::<omicron_nexus::Server>(
+        "test_oximeter_reregistration",
+    )
+    .await;
     let db = &context.database;
     let producer_id = nexus_test_utils::PRODUCER_UUID.parse().unwrap();
     let oximeter_id = nexus_test_utils::OXIMETER_UUID.parse().unwrap();
