+++ nexus/tests/integration_tests/rack.rs
@@ -4,10 +4,12 @@
 
 use nexus_test_utils::http_testing::AuthnMode;
 use nexus_test_utils::http_testing::NexusRequest;
-use nexus_test_utils::ControlPlaneTestContext;
 use nexus_test_utils_macros::nexus_test;
 use omicron_nexus::external_api::views::Rack;
 use omicron_nexus::TestInterfaces;
+
+type ControlPlaneTestContext =
+    nexus_test_utils::ControlPlaneTestContext<omicron_nexus::Server>;
 
 #[nexus_test]
 async fn test_list_own_rack(cptestctx: &ControlPlaneTestContext) {
