+++ nexus/tests/integration_tests/ssh_keys.rs
@@ -4,12 +4,14 @@ use http::{method::Method, StatusCode};
 
 use nexus_test_utils::http_testing::{AuthnMode, NexusRequest};
 use nexus_test_utils::resource_helpers::objects_list_page_authz;
-use nexus_test_utils::ControlPlaneTestContext;
 use nexus_test_utils_macros::nexus_test;
 
 use omicron_common::api::external::IdentityMetadataCreateParams;
 use omicron_nexus::external_api::params::SshKeyCreate;
 use omicron_nexus::external_api::views::SshKey;
+
+type ControlPlaneTestContext =
+    nexus_test_utils::ControlPlaneTestContext<omicron_nexus::Server>;
 
 // Note: we use UnprivilegedUser in this test because unlike most tests, all the
 // endpoints here _can_ be accessed by that user and we want to explicitly
