+++ nexus/tests/integration_tests/vpc_routers.rs
@@ -13,13 +13,15 @@ use nexus_test_utils::resource_helpers::objects_list_page_authz;
 use nexus_test_utils::resource_helpers::{
     create_organization, create_project, create_vpc,
 };
-use nexus_test_utils::ControlPlaneTestContext;
 use nexus_test_utils_macros::nexus_test;
 use omicron_common::api::external::IdentityMetadataCreateParams;
 use omicron_common::api::external::IdentityMetadataUpdateParams;
 use omicron_nexus::external_api::params;
 use omicron_nexus::external_api::views::VpcRouter;
 use omicron_nexus::external_api::views::VpcRouterKind;
+
+type ControlPlaneTestContext =
+    nexus_test_utils::ControlPlaneTestContext<omicron_nexus::Server>;
 
 #[nexus_test]
 async fn test_vpc_routers(cptestctx: &ControlPlaneTestContext) {
